requirements.txt
couchbase==4.2.1
langchain-community==0.0.37
langchain-openai==0.1.6
openai==1.26.0
pandas==2.2.2
python-dotenv==1.0.1
streamlit==1.34.0
tqdm==4.66.4
